Styling changes
- Some styling added to the event adder popup
- Links take you to /test now instead of /dashboard
- Event adding causing permissions error still.
