+++ frontend/remindit-frontend/src/index.css
@@ -7,6 +7,10 @@
     font-family: 'Montserrat';
 }
 
+:root {
+    color-scheme: dark;
+}
+
 body,
 html,
 #root {
@@ -19,4 +23,27 @@ html,
 
 html {
     background: #263238;
+}
+
+.MuiDialog-root.MuiModal-root .MuiPaper-root.MuiPaper-elevation.MuiPaper-rounded {
+    background: #222222;
+}
+
+.MuiDialog-root.MuiModal-root .MuiPaper-root.MuiPaper-elevation.MuiPaper-rounded .MuiTypography-root.MuiTypography-overline,
+.MuiPickersDay-root.MuiPickersDay-dayWithMargin,
+.MuiTypography-root.MuiTypography-caption.MuiDayCalendar-weekDayLabel,
+.MuiButtonBase-root.MuiIconButton-root.MuiIconButton-sizeMedium.MuiPickersArrowSwitcher-button,
+.MuiButtonBase-root.MuiIconButton-root.MuiIconButton-sizeSmall.MuiPickersCalendarHeader-switchViewButton,
+span.MuiTypography-root.MuiTypography-h3.MuiPickersToolbarText-root,
+span.MuiTypography-root.MuiTypography-subtitle2.MuiTimePickerToolbar-ampmLabel.MuiPickersToolbarText-root,
+span.MuiClockNumber-root {
+    color: #fff;
+}
+
+.MuiButtonBase-root.MuiPickersDay-root.MuiPickersDay-dayWithMargin.MuiPickersDay-today {
+    border: solid 2px white;
+}
+
+.MuiPaper-root.MuiMenu-paper.MuiPaper-elevation.MuiPaper-rounded.MuiPaper-elevation8.MuiPopover-paper {
+    background: #222222
 }
